configure 뷰페이지 Select 구현

+++ src/main/java/com/skhu/practice/Repository/AlbumRepository.java
@@ -3,5 +3,6 @@ package com.skhu.practice.Repository;
 import com.skhu.practice.Entity.Album;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlbumRepository extends JpaRepository<Album, Integer> {
+public interface AlbumRepository extends JpaRepository<Album, Long> {
+
 }
